moved redis config to seperate file
